A TON of changes. Read description.
Fixed the new commands having inconsistent command syntax.
Fixed AcceptChallenge using the challengers UUID instead of username.
Fixed the /CheckBadges and /GymList command GUI's not being the size of a double chest
Added a cap of 54 gyms on /AddGym
Fixed Tab Completion literally not working
FINALLY fixed the bug where you could not remove an offline leader from a gyms leader list
/AddLeader can now add NPC Leaders again
Standardized all of the incorrect usage messages and fixed descriptions in the /help command of all AGP commands.
Fixed leaderJoinMessage controlling whether a gym automatically opens or not when a leader joins.
Added an open gym message if openAnnouncement is true in the AGP announcements config
Added a prefix option for AGP in the announcements config which will control the AGP prefix globally
Removed the pastebin cacheing code for AGP Developers and Helpers and replaced it with a hard coded system.